containers/risk_warning_system/sentenced_object_detection/requirements.txt
# Florence-2 compatible stack (pin for reproducibility)
torch==2.3.1
transformers==4.49.0
huggingface-hub==0.26.2
Pillow==10.4.0
timm==1.0.11
einops==0.8.0
